GymForces/requirements.txt
django == 3.1.2
pillow == 8.0.1
django-crispy-forms
django-rest-framework
social-auth-app-django
django-pwa